Refactoring blogApiTests, adding one record ind import.sql
